(#77) refactor: Factory extension 파일단위로 분리
