+++ BoostRunClub/Coordinators/PrepareRunCoordinator.swift
@@ -19,9 +19,9 @@ final class PrepareRunCoordinator: BasicCoordinator, PrepareRunCoordinatorProtoc
 
     init(navigationController: UINavigationController, factory: PrepareRunSceneFactory = DependencyFactory.shared) {
         self.factory = factory
+        super.init(navigationController: navigationController)
         navigationController.view.backgroundColor = .systemBackground
         navigationController.setNavigationBarHidden(false, animated: true)
-        super.init(navigationController: navigationController)
     }
 
     override func start() {
