+++ BoostRunClub/Factory/Extensions/Factory+ActivityScene.swift
@@ -1,0 +1,15 @@
+//
+//  Factory+ActivityScene.swift
+//  BoostRunClub
+//
+//  Created by Imho Jang on 2020/12/06.
+//
+
+import UIKit
+
+protocol ActivitySceneFactory {
+    //    func makeActivityVC(with viewModel: ActivityViewModelTypes) -> UIViewController
+    //    func makeActivityVM() -> ActivityViewModelTypes
+}
+
+extension DependencyFactory: ActivitySceneFactory {}
