+++ BoostRunClub/Factory/Extensions/Factory+SplitScene.swift
@@ -1,0 +1,15 @@
+//
+//  Factory+SplitScene.swift
+//  BoostRunClub
+//
+//  Created by Imho Jang on 2020/12/06.
+//
+
+import UIKit
+
+protocol SplitSceneFactory {
+//    func makeSplitVC()
+//    func makeSplitVM()
+}
+
+extension DependencyFactory: SplitSceneFactory {}
